+++ app/Http/Controllers/ProductManufactureController.php
@@ -7,6 +7,8 @@ use Illuminate\Http\Request;
 use App\Http\Requests\ProductManufactureRequest;
 use App\Product;
 use App\Material;
+use Illuminate\Support\Facades\Auth;
+
 
 class ProductManufactureController extends Controller
 {
@@ -28,11 +30,14 @@ class ProductManufactureController extends Controller
      */
     public function create()
     {
-         
-        $products = Product::get();
-        return view('admin.productmanufactures.create',[
-            'products'=>$products,
+        if (Auth::user()->hasPermissionTo('add-product-manufacture')) {
+            $products = Product::get();
+            return view('admin.productmanufactures.create',[
+                'products'=>$products,
             ]);
+        }else {
+            abort(503);
+        }
     }
 
     /**
@@ -43,17 +48,21 @@ class ProductManufactureController extends Controller
      */
     public function store(ProductManufactureRequest  $request)
     {
-        $validated = $request->validated();
-        foreach ($validated['group'] as $material) {
-            $productmanufacture = new ProductManufacture;
-            $productmanufacture->product_id = $validated['product_id'];
-            $productmanufacture->material_id = $material['material_id'];
-            $productmanufacture->required_quantity = $material['required_quantity'];
-            $productmanufacture->waste_percentage = $material['waste_percentage'] / 100;
-            $productmanufacture->save();
+        if (Auth::user()->hasPermissionTo('add-product-manufacture')) {
+            $validated = $request->validated();
+            foreach ($validated['group'] as $material) {
+                $productmanufacture = new ProductManufacture;
+                $productmanufacture->product_id = $validated['product_id'];
+                $productmanufacture->material_id = $material['material_id'];
+                $productmanufacture->required_quantity = $material['required_quantity'];
+                $productmanufacture->waste_percentage = $material['waste_percentage'] / 100;
+                $productmanufacture->save();
+            }
+            $request->session()->flash('message',__('productmanufactures.massages.created_succesfully'));
+            return redirect(route('productmanufactures.index'));
+        }else {
+            abort(503);
         }
-        $request->session()->flash('message',__('productmanufactures.massages.created_succesfully'));
-        return redirect(route('productmanufactures.index'));
     }
 
     /**
@@ -75,13 +84,17 @@ class ProductManufactureController extends Controller
      */
     public function edit(ProductManufacture $productmanufacture)
     {
-        $materials = Material::get();
-        $products = Product::get();  
-        return view('admin.productmanufactures.edit',[
-            'productmanufacture'=>$productmanufacture,
-            'materials'=>$materials,
-            'products'=>$products,
-        ]);
+        if (Auth::user()->hasPermissionTo('edit-product-manufacture')) {
+            $materials = Material::get();
+            $products = Product::get();  
+            return view('admin.productmanufactures.edit',[
+                'productmanufacture'=>$productmanufacture,
+                'materials'=>$materials,
+                'products'=>$products,
+            ]);
+        }else {
+            abort(503);
+        }
     }
 
     /**
@@ -93,24 +106,29 @@ class ProductManufactureController extends Controller
      */
     public function update(Request $request, ProductManufacture $productmanufacture)
     {
-        $validated = $request->validate([
-            'material_id'=>'required',
-            'product_id'=>'required',
-            'required_quantity'=>'required',
-            'waste_percentage'=>'required',
-        ],[
-            'material_id'=>__('productmanufactures.material_id'),
-            'product_id'=>__('productmanufactures.product_id'),
-            'required_quantity'=>__('productmanufactures.required_quantity'),
-            'waste_percentage'=>__('productmanufactures.waste_percentage'),
-        ]);
-        $productmanufacture->product_id = $validated['product_id'];
-        $productmanufacture->material_id = $validated['material_id'];
-        $productmanufacture->required_quantity = $validated['required_quantity'];
-        $productmanufacture->waste_percentage = $validated['waste_percentage'];
-        $productmanufacture->save();
-        $request->session()->flash('message',__('productmanufactures.massages.updated_succesfully'));
-        return redirect(route('productmanufactures.index'));
+        if (Auth::user()->hasPermissionTo('edit-product-manufacture')) {
+            $validated = $request->validate([
+                'material_id'=>'required',
+                'product_id'=>'required',
+                'required_quantity'=>'required',
+                'waste_percentage'=>'required',
+            ],[
+                'material_id'=>__('productmanufactures.material_id'),
+                'product_id'=>__('productmanufactures.product_id'),
+                'required_quantity'=>__('productmanufactures.required_quantity'),
+                'waste_percentage'=>__('productmanufactures.waste_percentage'),
+            ]);
+            $productmanufacture->product_id = $validated['product_id'];
+            $productmanufacture->material_id = $validated['material_id'];
+            $productmanufacture->required_quantity = $validated['required_quantity'];
+            $productmanufacture->waste_percentage = $validated['waste_percentage'];
+            $productmanufacture->save();
+            $request->session()->flash('message',__('productmanufactures.massages.updated_succesfully'));
+            return redirect(route('productmanufactures.index'));
+        }else {
+            abort(503);
+        }
+        
     }
 
     /**
@@ -121,9 +139,14 @@ class ProductManufactureController extends Controller
      */
     public function destroy(Request $request,ProductManufacture $productmanufacture)
     {
-        $productmanufacture->delete();
-        $request->session()->flash('message',__('productmanufactures.massages.deleted_succesfully'));
-        return redirect(route('productmanufactures.index'));
+        if (Auth::user()->hasPermissionTo('delete-product-manufacture')) {
+            $productmanufacture->delete();
+            $request->session()->flash('message',__('productmanufactures.massages.deleted_succesfully'));
+            return redirect(route('productmanufactures.index'));
+        }else {
+            abort(503);
+        }
+        
     }
 
     public function material_select2_ajax (Request $request){
